refactored session handling with custom hook to fit ESLint best practice
